feat: Makes ConfigurationProperty Class not to use Bean / Interface
Fixes gh-4110
